在 X 型键盘视图中，支持图标绘制并调整布局样式

+++ app/src/main/java/org/crazydan/studio/app/ime/kuaizi/utils/ThemeUtils.java
@@ -73,7 +73,7 @@ public class ThemeUtils {
         }
 
         String[] splits = border.trim().split("\\s+");
-        int width = ScreenUtils.dpToPx(Integer.parseInt(splits[0]));
+        float width = ScreenUtils.dpToPx(Float.parseFloat(splits[0]));
         int color = Color.parseColor(splits[1]);
 
         if (width > 0) {
